+++ src/ia/corrections.py
@@ -381,7 +381,7 @@ def Chromatic_abbrevation_correction(im, correction_channel, target_channel='647
         single_im_size: full image size before any slicing, list of 3 (default:[30,2048,2048])
         drift: 3d drift of the image, which could be corrected at the same time, list/array of 3
         correction_folder: correction folder to find correction profile, string of path (default: Z://Corrections/)
-        profile_dtype: data type for correction profile, numpy datatype (default: np.float)
+        profile_dtype: data type for correction profile, numpy datatype (default: np.float32)
         image_dtype: image data type, numpy datatype (default: np.uint16)
         cc_profile_name: chromatic correction file basename, str (default: 'chromatic_correction')
         verbose: say something!, bool (default: True)"""
@@ -541,7 +541,7 @@ def _mean_xy_profle(im_filename, color, all_colors=_allowed_colors, frame_per_co
                                z_shift_corr=z_shift_corr, hot_pixel_remove=hot_pixel_remove, 
                                illumination_corr=False, chromatic_corr=False,
                                return_limits=False, verbose=False)
-    _im = _im.astype(np.float)
+    _im = _im.astype(np.float32)
     # seeding
     _seeds = visual_tools.get_seed_in_distance(_im, th_seed_percentile=seeding_th_per, dynamic=True)
     for _sd in _seeds:
@@ -1481,7 +1481,7 @@ def Generate_bleedthrough_correction(folder_list, channel_list,
 
         # transpose first two axes
         bld_corr_profile = bld_corr_profile.transpose((1,0,2,3))
-        inv_corr_profile = np.zeros(np.shape(bld_corr_profile), dtype=np.float)
+        inv_corr_profile = np.zeros(np.shape(bld_corr_profile), dtype=np.float32)
         for _i in range(np.shape(bld_corr_profile)[-2]):
             for _j in range(np.shape(bld_corr_profile)[-1]):
                 inv_corr_profile[:,:,_i,_j] = np.linalg.inv(bld_corr_profile[:,:,_i,_j])
@@ -1516,7 +1516,7 @@ def Bleedthrough_correction(input_im, crop_limits=None, all_channels=_allowed_co
         drift: 3d drift vector for this image, 1d-array (default:np.array([0,0,0]))
         correction_folder: correction folder to find correction profile, string of path (default: Z://Corrections/)
         profile_basename: base filename for bleedthrough correction profile file, str (default: 'Bleedthrough_correction_matrix')
-        profile_dtype: data type for correction profile, numpy datatype (default: np.float)
+        profile_dtype: data type for correction profile, numpy datatype (default: np.float32)
         image_dtype: image data type, numpy datatype (default: np.uint16)
         return_limits: return modified limits, 3x2 np.ndarray
         verbose: say something!, bool (default: True)
@@ -1576,7 +1576,7 @@ def Bleedthrough_correction(input_im, crop_limits=None, all_channels=_allowed_co
             print(f"-- correcting bleedthrough for images")
         _ims = input_im
         _dft_limits = crop_limits
-    _ims = [_im.astype(np.float) for _im in _ims]
+    _ims = [_im.astype(np.float32) for _im in _ims]
     #print(time.time()-_start_time)
     # load profile
     if verbose:
